Add files via upload
HTML: It is a simple file having the basic structure of the webpage and ID for the clock's body and for the second, minute, and hour hands. 
CSS: The CSS is used just for making the clock actually look a bit nicer. We have basically centered our clock in the middle of the webpage.